requirements.txt
django_extensions
Django==3.1
djangorestframework==3.12.2
model_bakery==1.2.0

# heroku files
dj-database-url
psycopg2
gunicorn
whitenoise